Changed default configuration file handling.
Instead of loading defaults and then saving a generated copy of the
configuration, the default configuration file is just written directly
if a user configuration file doesn't already exist.

How blocks are broken based on whether an enchanted tool will affect all
blocks or not was refactored into a function.

Documentation was also added to all source files.